ci-build: More fixes

+++ docker/ci-build.sh
@@ -82,12 +82,14 @@ add_to_repository() {
 	chmod 400 ~/.ssh/id_rsa
 
 	REPO_BRANCH="${REPO_BRANCH:main}"
-	ARTIFACTS=$(ls ${PACKAGE_ROOT}/../*.{dsc,deb,orig*,debian*,changes})
-
-	rsync -avzp -e \
-		"ssh -o StrictHostKeyChecking=no -p ${DEPLOY_PORT}" \
-		"${ARTIFACTS}" \
-		"${DEPLOY_ACCOUNT}:/var/opt/repo-debpm-incoming/${REPO_BRANCH}"
+	for file in ${PACKAGE_ROOT}/../*.{dsc,deb,orig*,debian*,changes}; do
+		if [ -f $file ]; then
+			rsync -avzp -e \
+				"ssh -o StrictHostKeyChecking=no -p ${DEPLOY_PORT}" \
+				"${file}" \
+				"${DEPLOY_ACCOUNT}:/var/opt/repo-debpm-incoming/${REPO_BRANCH}"
+		fi
+	done
 }
